+++ src/OpenTelemetry.Exporter.Prometheus.AspNetCore/CHANGELOG.md
@@ -62,6 +62,10 @@ Notes](../../RELEASENOTES.md).
   specification.
   ([#7266](https://github.com/open-telemetry/opentelemetry-dotnet/issues/7266))
 
+* Abort scrape request processing if request exceeds the value specified by the
+  `X-Prometheus-Scrape-Timeout-Seconds` HTTP request header.
+  ([#7252](https://github.com/open-telemetry/opentelemetry-dotnet/issues/7252))
+
 * Export `{name}_created` series for counters and histograms when using
   OpenMetrics and a start time is available.
   ([#7223](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7223))